requirements.txt
boto3
awscli-local
psycopg2-binary